refactor: split log emission to encrypt and a log, remove address input (#6987)
Removes the `contract_address` input from the context emission functions
as it is always the contract itself if honest, so can just as well be
auto populated for less error.

Split the log emission into two variants, namely one that is just
emitting whatever passed, and another that is encrypting first. This is
to make it straight forward to encrypt with a different scheme if
desired.