fix(elaborator): Fix duplicate methods error (#5225)
# Description

## Problem\*

Resolves "duplicate method" error when declaring trait methods on
different traits with the same method name in the elaborator.

## Summary\*



## Additional Context

This was the last error type in aztec packages for the elaborator.
~~After this, there are still some extra unused variable warnings I will
fix but the entire crate tests successfully.~~ Nevermind, looks like
aztec-nr just naturally has many unused variables

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.