fix(elaborator): Fix regression introduced by lazy-global changes (#5223)
# Description

## Problem\*

Fixes a regression introduced by #5191

## Summary\*

When a global is lazily evaluated, the local module id and file id of
the elaborator would change without being reset to the previous value.
This lead to scoping issues after the global was evaluated.

## Additional Context

I was unable to create a repro for this aside from running `nargo t
--use-elaborator` on `aztec-nr` unfortunately.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.