fix: add support for nested arrays returned by oracles (#5132)
# Description

## Problem\*

Resolves #4561

## Summary\*
Generates brillig arrays of arrays during brillig-gen for types having
nested arrays.
Populate these nested brillig arrays in the brillig VM when data
returned from oracles does not match the size of the outer array


## Additional Context
For this to work, it is required that the oracle returns flatten values,
although its Noir signature is a nested type.
It's certainly possible to deserialise nested json arrays but I could
not get it done easily so it'll be better handled in a separate PR.


## Documentation\*

Check one:
- [ ] No documentation needed.
- [X] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Álvaro Rodríguez <sirasistant@gmail.com>
Co-authored-by: TomAFrench <tom@tomfren.ch>