chore: automate necessary changes to outward sync into noir-lang/noir (#7049)
This PR automates syncs out to noir-lang/noir similarly to how it was
done for inward syncs.

The main difference is that we need to save the version of `bb` which
noir-lang/noir is using so we can restore it on any outward sync. We
then save the version of bb.js last seen on every inbound sync into the
`bb-version` file and reapply it on outbound syncs.