feat: implement comptime support for `array_len` builtin (#5272)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR implements support for the `array_len` builtin so we can query
the length of arrays in `comptime` blocks. Ideally we'd tie this closer
to the `Intrinsics` struct in `noirc_evaluator` but we can handle this
morn cleanly later.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.