chore: create separate crate just for noir artifacts (#5162)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR splits off the definitions for Noir's build artifacts from the
`nargo` crate which adds a bunch of high-level logic. This makes it
easier to pull in and work on these types without having to pull in the
whole compiler.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.