feat(macros): Add `TopLevelItem` and `Type` types for quoting top-level definitions and types (#5267)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Adds a `TopLevelItem` and `Type` keywords & types for quoting type
definitions, functions, trait impls, etc. and types respectively. The
former is needed to identify items which can be unquoted at the
top-level of a file - e.g. to prevent random expressions from being
unquoted at the top level and to prevent trait impls from being unquoted
in the middle of a function.

## Additional Context

There is no way currently to create a TopLevelItem. The intended method
is to be able to specify the type you want to quote in a quote
expression: `quote TopLevelItem { ... }` but our current parser makes
this very difficult since the top level statement parser would need to
become recursive and be passed down through each function. For now I'm
planning to have compiler API function's as a work around - e.g:

```rs
impl TopLevelItem {
    #[builtin(new_trait_impl)]
    pub fn new_trait_impl(generics: Vec<String>, trait: Type, typ: Type, methods: Vec<TopLevelItem>) -> TopLevelItem { }
}
```

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.