fix: avoid unnecessarily splitting expressions with multiplication terms with a shared term (#5291)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR addresses an issue where we were unnecessarily splitting an
expression based on an example which Zac found while working on
`noir-edwards`. We were being overly restrictive and only accepting the
case where both witnesses in the multiplication are shared with other
terms

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.