fix: Add more thorough check for whether a type is valid when passing it from constrained code to unconstrained code (#5009)
# Description

## Problem\*

Resolves #5008 

## Summary\*

Thank you to @nventuro for the minimal example and @TomAFrench for
looking into this. As a result it was a quick fix.

## Additional Context

All type system fixes must currently be duplicated in the elaborator so
I've included this there as well.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: TomAFrench <tom@tomfren.ch>
Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>