chore: refactor test case generation in build.rs (#5280)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

The `build.rs` file for generating the test cases is getting a bit
unmanagable imo. This PR factors out some repeated code and adds a new
helper function so we just need to pass in the body of the test.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.