fix: update `in_contract` flag before handling function metadata in elaborator (#5292)
…

# Description

## Problem\*

Resolves #5288 

## Summary\*

This PR fixes #5288 by updating `self.in_contract` before we define
function metadata to match `self.local_module`. This is mildly hacky and
we may want to have a nicer "switch modules and update everything which
depends on it" helper function in future.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.