feat(tracer): Count lines, not opcodes
Before this commit, a tracing step was a jump over one opcode.

After this commit, a tracing step is a jump over one line in the source
code.

The test relevant to tracing is updated, so now the expected number of
tracing steps is less (7, instead of 14). For some reason, the test does
not print "warning: no call stack" anymore. I'm not completely sure why,
but this is a wanted improvement, so the expectation for that message is
removed.

There is a potential low-priority bug, involving function calls. A TODO
for it is added in the code.

**Details**

The following functions are no longer private in `DebugContext`:
 * `get_call_stack`
 * `get_source_location_for_opcode_location`

Although this is not ideal, it is an acceptable middle-term solution,
until common logic of the debugger and the tracer is extracted out of
the debugger.

The following functions are added to `DebugContext`, in order to surface
functionality of the contained `debug_artifact`:
 * `get_filepath_for_location`
 * `get_line_for_location`

In `tracer/src/lib.rs`:
 * the initializing logic that checks that the circuit contains opcodes
   is now moved into the `trace_circuit` function, rather than in the
   constructor for `TracingContext`;
 * `debug_context` is renamed to `tracing_context`, which matches the
   type;
 * the error for when there is a breakpoint during tracing is promoted
   to `panic!`, as it should never happen (indicates a serious bug if it
   happens);
 * `last_result` is no longer stored in TracingContext, to keep things
   purer.

**Next steps**

More tests need to be added to explore the error modes. When is there no
call stack? When is there no mapping from opcodes to source locations?
This is not a priority, however, so the first follow-up that is coming
is actually recording information on each step and saving it in the
tracing artifact.