feat: Implement `slice_push_back` in the comptime interpreter (#5312)
# Description

## Problem\*

## Summary\*

Implements `slice_push_back` in the comptime interpreter. This was
needed in the vertical slice example to build a slice of quoted fields
for the type in the impl.

## Additional Context

With this a slightly modified version of the vertical slice code now
runs. Notably:
- Generic functions do not work, so a `map` function like in the
original cannot be used
- An annotation function can return a quoted impl, but the impl isn't
actually inserted into the program yet. So the derived impl does nothing
currently.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.