fix: skip emission of brillig calls which will never be executed (#5314)
# Description

## Problem\*

Pulls across fix made in sync PR at
https://github.com/AztecProtocol/aztec-packages/pull/7134/commits/e39be3dc0ee89d9a246366a1f88eb36b26f25751

## Summary\*

For some reason in the `noir-contracts` workspace, we're emitting a
_lot_ (100+ MB worth) of brillig call opcodes which have their predicate
hardcoded to zero (i.e. they will never actually be executed).

This PR adds a check before we emit a brillig call opcode to see if the
predicate is zero and, if so, we skip it and zero out all the outputs at
compile time.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.