chore: add back Pedersen blackbox functions (revert PR 5221) (#5318)
# Description

## Problem\*

This PR adds (back) the blackbox functions for Pedersen hash/commitment,
until we get a similar circuit (in term of number of gates) with the
Noir implementation.

## Summary\*
The Noir version is kept and assert to ouput the same value in a test.


## Additional Context
This PR reverts the PR #5221


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>
Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>