feat(stdlib): Update stdlib to use explicit numeric generics (#5306)
# Description

## Problem\*

Builds upon changes in https://github.com/noir-lang/noir/pull/5155 as to
avoid warnings in the stdlib.

## Summary\*

This PR simply switches away from all usages of implicit numeric
generics in the stdlib as to avoid the warning created in #5155.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>