fix: don't benchmark the "prove" command as it doesn't exist anymore (#5323)
# Description

## Problem

Related to #4754

## Summary

In order to improve execution performance we first need to benchmark
it... but, the current benchmark code also benchmarks the "prove"
command which doesn't exist anymore, so it can't run. This PR remove the
benchmark command related to the "prove" command.

I considered extracting the strings "prove" and "command" to constants,
so that if they are removed we would get a compile error, but... maybe
chances of a command getting removed are very low, so probably not worth
it.

## Additional Context

None.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.