chore: remove `is_unconstrained_fn` field from elaborator (#5335)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

In the vein of #5324, we now just look up whether the `current_function`
is unconstrained rather than tracking this on the elaborator.

This does come with a little bit of a perf degradation as we need to
look up the function on each jump but that should be pretty negligible.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.