feat(trace): simplify trace test scaffold
This commit simplifies the testing scaffold (in
`tooling/nargo_cli/build.rs`) because instead of comparing a message
about the number of steps, we now compare the full traces.

It also updates `expected_trace.json` in the test `2_function_calls`,
but the change is not related to code change in this commit. Rather, the
test was broken in the `plonky2` (default) branch, and this PR fixes
that as a drive-by.

(cherry picked from commit 59f6c6993b35d92ff6d8068add0b52467080f2b9)