feat(trace): fix bug of missing steps
If the next source line for `nargo debug` is a function call, one
debugger step includes it, together with the first line of the called
function. This is just how `nargo debug` works and a fact of life we
choose not to change.

Instead, this commit takes this into account in order to produce a
correct trace that includes as steps both the invocation site, as well as the
first line of the function.

The tests are updated to demonstrate the effect of the bugfix. Note the
new lines that show up in the expected traces now.

As a drive-by, the initial location that is registered now depends on
the implementation of create_unknown, rather than an ad-hoc invalid
location. (See change in expected_trace.json for 1_mul and call to
`tracer.start`.)

Run tests with:

```
cargo test trace
```

Note: the change in `get_current_source_locations` is actually simpler
than it looks: new scopes and indentation are introduced, however, which
trips up the change diff algorithm.