feat: TXE fixes to avm opcodes and missing oracles, forced ci failure (#7252)
Makes CI fail in case TXE test fail. Noticed they were broken, but
master was still green and removed some unnecessary commands in the
docker build.

Also reverted some of the changes from
https://github.com/AztecProtocol/aztec-packages/pull/7237 since we need
the length of the data we're about to read in `avmOpcodeStorageRead`
(the oracle cannot obtain this information from the assigned output
variable.