fix: correctly detect signed/unsigned integer overflows/underflows (#5375)
# Description

## Problem

Resolves #5372

## Summary

Correctly detects overflows/underflows for basic integer types. The sign
of a field element value is now also propagated to the ssa phase. But
previously, the field value was applied some casting before reaching the
ssa phase so checking that value like that was a bit strange. In this PR
that conversion is delayed until the ssa phase, and the value can be
checked prior to that conversion.

## Additional Context

None.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.