feat: lsp "find all references" (#5395)
# Description

## Problem

Resolves #5393

## Summary

Implements "Find all references". The logic for it is the same as the
one for "Rename symbol" so it was straight-forward to do.


https://github.com/noir-lang/noir/assets/209371/5eef03bf-6170-48f3-beaa-87dcc86532d2

## Additional Context

In the video above, "Find all references" misses a few locations. I
captured that as a separate issue ( #5394 ) because since both features
rely on the same logic, fixing that would make this one work too... so
this PR is mostly about enabling "Find all references" rather than the
logic to actually find all references.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>