test(unconstrained): add simple test
The purpose of this test is to show a minimal useful program that makes
use of the `unconstrained` capabilities of Noir and to show that the
`nargo` tool is capable of executing it and performing the associated
assert.

Note that the `unconstrained` code is, unsurprisingly, not included in
the constraint system, so it is not part of the ZK proof/verification.
It is just a utility that allows programs to get around performance
limitations of the arithmetic circuit paradigm. Modern zkVMs provide a
way to include such code in the proof as well, but these are in even
earlier stage of development at the moment.

To run test:

```
cargo test simple_unconstrained
```