fix: lsp find struct reference in return locations and paths (#5404)
# Description

## Problem

Resolves #5394

## Summary

Finds struct locations in return types and inside paths.

## Additional Context

Making it work for the return type was simple: the code to register it
is now done a bit earlier so it works in all cases where a type is
referred as the last segment in a Path.

Making it work when the type is in the middle of a Path (like in
`foo::Foo::new`) was a bit trickier: the existing Path lookup only cared
about (and registered) the result of the last segment. So now there's
code to register each segment in turn... but this is only done if
running in LSP mode. So in `foo::Foo::new`, `foo` is linked to a module
and `Foo` is linked to a struct. The existing code relied on the
`DependencyId` for this, but a "module" dependency didn't exist... to
avoid introducing new variants that are only used in this reference
tracking I instead created a new enum, `ReferenceId`.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.