feat: Detect subgraphs that are completely independent from inputs or outputs (#5402)
# Description

## Problem\*

Currently a developer can create a circuit that would use unconstrained
values in such a way, that a part of the circuit becomes completely
disconnected from the circuit inputs or outputs, making this part
completely useless as it can be taken out of any proof. This PR
introduces an Ssa pass that informs the user of this issue.

## Summary\*

An Ssa pass that detects some underconstraining issues

![image](https://github.com/noir-lang/noir/assets/4798775/cb5849d2-715b-4a6c-9947-575c2938f37a)


## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.