chore: merge `BarretenbergVerifierBackend` and `BarretenbergBackend` (#5399)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

As pointed out by Palla in the bb.js doc, we're exporting two different
verifiers here. I'm not sure why this is the case so I've folded
`BarretenbergVerifierBackend` into `BarretenbergBackend` so
`BarretenbergVerifier` is the only `VerifierBackend` that doesn't also
implement `Backend`.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.