fix: lsp struct rename/reference difference (#5411)
# Description

## Problem

Resolves #5412

## Summary

When a function parameter is `self` and it refers to a struct type, that
variable is linked to that type (through a synthetic "Self" type):

```rust
struct Foo {
}

impl Foo {
  fn foo(self) { // <-- here
  }
}
```

Then, when renaming `Foo`, those `self` occurrences were incorrectly
renamed. To fix that, in this PR we don't track these synthetic `Self`
types in the reference graph.

The second thing here is that when using `Self` to refer to a struct
type, that `Self` was also incorrectly renamed. To fix this we now track
whether a reference to a type came from `Self` or from the actual struct
name. Then we don't rename these `Self` references.

Note that we still keep these `Self` references in order for "find all
references" to return them (this is similar to how rust-analyzer works).


https://github.com/noir-lang/noir/assets/209371/8eeeebd7-816f-4e8f-a6cc-426a5172c1a7

## Additional Context

None.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.