feat: lsp rename/find-all-references for type aliases (#5414)
# Description

## Problem

Resolves #5358

## Summary

Based on top of #5411

Implements the usual rename/find-all-references but this time for type
aliases.

I noticed some things related to name locations could be simplified (not
stored in a duplicate way) so there's also a small simplification here.


https://github.com/noir-lang/noir/assets/209371/77508029-7002-4708-9cc1-24fc879630db

## Additional Context

None.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.