feat: remove redundant `EnableSideEffects` instructions (#5440)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We currently pick up quite a lot of duplicated `EnableSideEffects`
instructions as we perform optimizations on the SSA which tends to
clutter up the output. This PR adds a small change to remember the
currently active predicate condition and skips any instructions which
will just will just reactivate the current predicate.

I've added a test case which shows the sort of SSA which we want to
avoid.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.