feat(optimization): Deduplicate more instructions (#5457)
# Description

## Problem\*

## Summary\*

@TomAFrench's recent SSA changes got me thinking more about the passes
and I was looking at how we deduplicated instructions. We never
deduplicated instructions which can require a predicate so I've added
the ability to do so here.

We only deduplicate these instructions after flattening and to do so we
cache the SideEffectsEnabled var in addition to the instruction - and
only if the instruction actually relies on the variable.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.