feat: prefix operator overload trait dispatch (#5423)
# Description

## Problem

Resolves #5000

## Summary

Allows `impl Not` and `impl Neg` to be picked up at runtime.

Note that this doesn't work yet at comptime: I'd prefer to do it in a
separate PR (mainly to keep PR smaller, but also because I'm not sure
how to test it).

## Additional Context

None.

## Documentation

Is documentation needed for this?

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>